refactor: change ActivityPanel folder name
