+++ src/components/ActivityPanel/index.tsx
