+++ src/layouts/MainLayout/index.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { Outlet } from 'react-router-dom';
 
+import { ActivityPanelToggle } from '@/components/ActivityPanel/Toggle';
 import { Header } from '@/components/Header';
 import { Sidebar } from '@/components/Sidebar';
 import { RouteWrapper } from '@/routes/RouteWrapper';
@@ -18,6 +19,7 @@ export const MainLayout = () => {
 
           <main className="h-screen">
             <Outlet />
+            <ActivityPanelToggle />
           </main>
         </div>
       </div>
